Change algorithm to get rid of rearrangeSegment
